# memory/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for CerberOS memory, chat, personal info, system event, vault, and agent execution services.",
        "title": "Memory Service API",
        "contact": {},
        "version": "v1"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/orchestrator/records": {
            "get": {
                "description": "Retrieves orchestrator records by data_type and task/orchestrator filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator"
                ],
                "summary": "Query orchestrator records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Data type",
                        "name": "data_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Orchestrator task ref",
                        "name": "orchestrator_task_ref",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Inclusive lower bound timestamp",
                        "name": "from_timestamp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Inclusive upper bound timestamp",
                        "name": "to_timestamp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Use not_terminal to exclude terminal task states",
                        "name": "state_filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Creates or upserts one orchestrator record depending on data_type semantics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator"
                ],
                "summary": "Write orchestrator record",
                "parameters": [
                    {
                        "description": "Orchestrator record payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.WriteOrchestratorRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/orchestrator/records/latest": {
            "get": {
                "description": "Retrieves the latest orchestrator record for a task and data_type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator"
                ],
                "summary": "Read latest orchestrator record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Data type",
                        "name": "data_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.WriteOrchestratorRecordRequest": {
            "type": "object",
            "properties": {
                "data_type": {
                    "type": "string"
                },
                "orchestrator_task_ref": {
                    "type": "string"
                },
                "payload": {
                    "type": "object"
                },
                "plan_id": {
                    "type": "string"
                },
                "subtask_id": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                },
                "ttl_seconds": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Internal-API-Key",
            "in": "header"
        }
    }
}